created cards.js file
