+++ src/index.js
@@ -1,45 +1,11 @@
-import addLike from './modules/likes.js';
 import './style.css';
+import pokemonCards from './modules/cards';
+import addLike from './modules/likes.js';
 
 const BASE_API_URL = 'https://pokeapi.co/api/v2';
 
 const itemContainer = document.getElementById('itemContainer');
 const itemCounter = document.getElementById('itemCounter');
-const pokemonCards = [
-  { name: 'bulbasaur', id: 1, description: 'A Grass/Poison-type Pokémon with a plant bulb on its back.' },
-  { name: 'ivysaur', id: 2, description: 'The evolved form of Bulbasaur, known for the large flower on its back.' },
-  { name: 'venusaur', id: 3, description: 'The final evolution of Bulbasaur, a powerful Grass/Poison-type Pokémon.' },
-  { name: 'squirtle', id: 7, description: 'A Water-type Pokémon that hides in its shell for protection.' },
-  { name: 'wartortle', id: 8, description: 'The evolved form of Squirtle, a Water-type Pokémon with impressive water cannons.' },
-  { name: 'blastoise', id: 9, description: 'The final evolution of Squirtle, a Water-type Pokémon with powerful water cannons on its back.' },
-  { name: 'charmander', id: 4, description: 'A Fire-type Pokémon known for the flame on the tip of its tail.' },
-  { name: 'charmeleon', id: 5, description: 'The evolved form of Charmander, a Fire-type Pokémon with a fiery personality.' },
-  { name: 'charizard', id: 6, description: 'A powerful Fire/Flying-type Pokémon with a fiery breath.' },
-  { name: 'pidgey', id: 16, description: 'A Normal/Flying-type Pokémon that is a common sight in many areas.' },
-  { name: 'pidgeotto', id: 17, description: 'The evolved form of Pidgey, a larger and more powerful Normal/Flying-type Pokémon.' },
-  { name: 'pidgeot', id: 18, description: 'The final evolution of Pidgey, a majestic Normal/Flying-type Pokémon.' },
-  { name: 'eevee', id: 133, description: 'A Normal-type Pokémon with multiple evolutionary paths.' },
-  { name: 'pikachu', id: 25, description: 'An Electric-type Pokémon known for its lightning bolt-shaped tail.' },
-  { name: 'charizard', id: 6, description: 'A powerful Fire/Flying-type Pokémon with a fiery breath.' },
-  { name: 'squirtle', id: 7, description: 'A Water-type Pokémon that hides in its shell for protection.' },
-  { name: 'bulbasaur', id: 1, description: 'A Grass/Poison-type Pokémon with a plant bulb on its back.' },
-  { name: 'eevee', id: 133, description: 'A Normal-type Pokémon with multiple evolutionary paths.' },
-  { name: 'jigglypuff', id: 39, description: 'A Normal/Fairy-type Pokémon that puts its foes to sleep by singing.' },
-  { name: 'meowth', id: 52, description: 'A Normal-type Pokémon known for its mischievous behavior and coin-collecting.' },
-  { name: 'magikarp', id: 129, description: 'A weak Water-type Pokémon that can evolve into a powerful Gyarados.' },
-  { name: 'raichu', id: 26, description: 'The evolved form of Pikachu, Raichu is an Electric-type Pokémon.' },
-  { name: 'mewtwo', id: 150, description: 'A Psychic-type Legendary Pokémon with immense power.' },
-  { name: 'vaporeon', id: 134, description: 'A Water-type Eeveelution known for its aquatic abilities.' },
-  { name: 'gyarados', id: 130, description: 'A powerful Water/Flying-type Pokémon known for its rage.' },
-  { name: 'snorlax', id: 143, description: 'A Normal-type Pokémon known for its sleeping habits and huge size.' },
-  { name: 'eevee', id: 133, description: 'A Normal-type Pokémon with multiple evolutionary paths.' },
-  { name: 'dragonite', id: 149, description: 'A powerful Dragon/Flying-type Pokémon with a kind heart.' },
-  { name: 'lapras', id: 131, description: 'A Water/Ice-type Pokémon known for its soothing song and serenity.' },
-  { name: 'machamp', id: 68, description: 'A Fighting-type Pokémon with impressive physical strength and power.' },
-  { name: 'meowth', id: 52, description: 'A Normal-type Pokémon known for its mischievous behavior and coin-collecting.' },
-  { name: 'magikarp', id: 129, description: 'A weak Water-type Pokémon that can evolve into a powerful Gyarados.' },
-  { name: 'raichu', id: 26, description: 'The evolved form of Pikachu, Raichu is an Electric-type Pokémon.' },
-];
 
 const openPopup = async (pokemon) => {
   try {
