+++ src/modules/cards.js
@@ -1,0 +1,13 @@
+const pokemonCards = [
+  { name: 'bulbasaur', id: 1, likes: 0, description: 'A Grass/Poison-type Pokémon with a plant bulb on its back.' },
+  { name: 'ivysaur', id: 2, description: 'The evolved form of Bulbasaur, known for the large flower on its back.' },
+  { name: 'venusaur', id: 3, description: 'The final evolution of Bulbasaur, a powerful Grass/Poison-type Pokémon.' },
+  { name: 'charmander', id: 4, description: 'A Fire-type Pokémon with a fiery personality.' },
+  { name: 'charmeleon', id: 5, description: 'The evolved form of Charmander, a Fire-type Pokémon with a fiery personality.' },
+  { name: 'charizard', id: 6, description: 'A powerful Fire/Flying-type Pokémon with a fiery breath.' },
+  { name: 'squirtle', id: 7, description: 'A Water-type Pokémon that hides in its shell for protection.' },
+  { name: 'wartortle', id: 8, description: 'The evolved form of Squirtle, a Water-type Pokémon with impressive water cannons.' },
+  { name: 'blastoise', id: 9, description: 'The final evolution of Squirtle, a Water-type Pokémon with powerful water cannons on its back.' },
+];
+
+export default pokemonCards;
