+++ src/modules/likes.js
@@ -1,5 +1,15 @@
 import { baseURL, appId } from './involveapi.js';
 
+const requestOptions = {
+  method: 'GET',
+  redirect: 'follow'
+};
+
+fetch("https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/sfozB4S7vDGGPXQ2BqJy/likes", requestOptions)
+  .then(response => response.text())
+  .then(result => console.log(result))
+  .catch(error => console.log('error', error));
+
 // add a like to the API
 const addLike = async (id) => {
   const url = `${baseURL}apps/${appId}/likes/`;
